requirements.txt
torch 
torchvision 
numpy 
pillow 
matplotlib 
streamlit 
pytest
httpx
zenml[server] 
zenml[pytorch]
huggingface_hub
lpips
scikit-image
scipy
seaborn
ipykernel   
opencv-python
scikit-learn
boto3